requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
python-dotenv==1.0.1
openai==1.12.0
langchain==0.1.4
langchain-openai==0.0.5
jinja2==3.1.3
json5==0.9.14
flask-cors>=5.0.0
requests>=2.31.0